Broke up functions
I broke up all and refactored the functions from the etl index into different files.
